chore: add script to join testnet with binaries (#129)
* chore: add script to join testnet with binaries

This commit introduces two new files. join-testnet-with-binaries.sh mirrors the joinTestnet.sh script but brings up a node and joins testnet with the binary. It does not depend on docker. launch-validator-with-binaries.sh mirrors launchValidator.sh and sets up the vald process and the tofnd process. You can run these scripts one after the other to bring up a node and set it up as a validator. You will still have to obtain tokens and stake them.